Move shared functions/variables to kzg_tests.py
