+++ tests/generators/kzg_4844/main.py
@@ -1,118 +1,22 @@
 """
-KZG 4844 test vectors generator
+KZG test vectors generator for EIP-4844
 """
 
-from hashlib import sha256
 from typing import Tuple, Iterable, Any, Callable, Dict
 
-from eth_utils import (
-    encode_hex,
-    int_to_big_endian,
-)
+from eth_utils import encode_hex
 
-from eth2spec.utils import bls
+from eth2spec.deneb import spec
+from eth2spec.gen_helpers.gen_base import gen_runner, gen_typing
 from eth2spec.test.helpers.constants import DENEB
 from eth2spec.test.helpers.typing import SpecForkName
-from eth2spec.gen_helpers.gen_base import gen_runner, gen_typing
-from eth2spec.deneb import spec
+from eth2spec.test.utils.kzg_tests import *
+from eth2spec.utils import bls
 
 
-def expect_exception(func, *args):
-    try:
-        func(*args)
-    except Exception:
-        pass
-    else:
-        raise Exception("should have raised exception")
-
-
-def field_element_bytes(x):
-    return int.to_bytes(x % spec.BLS_MODULUS, 32, spec.KZG_ENDIANNESS)
-
-
-def field_element_bytes_unchecked(x):
-    return int.to_bytes(x, 32, spec.KZG_ENDIANNESS)
-
-
-def encode_hex_list(a):
-    return [encode_hex(x) for x in a]
-
-
-def bls_add_one(x):
-    """
-    Adds "one" (actually bls.G1()) to a compressed group element.
-    Useful to compute definitely incorrect proofs.
-    """
-    return bls.G1_to_bytes48(
-        bls.add(bls.bytes48_to_G1(x), bls.G1())
-    )
-
-
-def evaluate_blob_at(blob, z):
-    return field_element_bytes(
-        spec.evaluate_polynomial_in_evaluation_form(spec.blob_to_polynomial(blob), spec.bytes_to_bls_field(z))
-    )
-
-
-BLS_MODULUS_BYTES = spec.BLS_MODULUS.to_bytes(32, spec.KZG_ENDIANNESS)
-
-G1 = bls.G1_to_bytes48(bls.G1())
-G1_INVALID_TOO_FEW_BYTES = G1[:-1]
-G1_INVALID_TOO_MANY_BYTES = G1 + b"\x00"
-G1_INVALID_P1_NOT_IN_G1 = bytes.fromhex("8123456789abcdef0123456789abcdef0123456789abcdef" +
-                                        "0123456789abcdef0123456789abcdef0123456789abcdef")
-G1_INVALID_P1_NOT_ON_CURVE = bytes.fromhex("8123456789abcdef0123456789abcdef0123456789abcdef" +
-                                           "0123456789abcdef0123456789abcdef0123456789abcde0")
-INVALID_G1_POINTS = [G1_INVALID_TOO_FEW_BYTES, G1_INVALID_TOO_MANY_BYTES,
-                     G1_INVALID_P1_NOT_IN_G1, G1_INVALID_P1_NOT_ON_CURVE]
-
-BLOB_ALL_ZEROS = spec.Blob()
-BLOB_ALL_TWOS = spec.Blob(b''.join([field_element_bytes(2) for n in range(4096)]))
-BLOB_RANDOM_VALID1 = spec.Blob(b''.join([field_element_bytes(pow(2, n + 256, spec.BLS_MODULUS)) for n in range(4096)]))
-BLOB_RANDOM_VALID2 = spec.Blob(b''.join([field_element_bytes(pow(3, n + 256, spec.BLS_MODULUS)) for n in range(4096)]))
-BLOB_RANDOM_VALID3 = spec.Blob(b''.join([field_element_bytes(pow(5, n + 256, spec.BLS_MODULUS)) for n in range(4096)]))
-BLOB_ALL_MODULUS_MINUS_ONE = spec.Blob(b''.join([field_element_bytes(spec.BLS_MODULUS - 1) for n in range(4096)]))
-BLOB_ALMOST_ZERO = spec.Blob(b''.join([field_element_bytes(1 if n == 3211 else 0) for n in range(4096)]))
-BLOB_INVALID = spec.Blob(b'\xFF' * 4096 * 32)
-BLOB_INVALID_CLOSE = spec.Blob(b''.join(
-    [BLS_MODULUS_BYTES if n == 2111 else field_element_bytes(0) for n in range(4096)]
-))
-BLOB_INVALID_LENGTH_PLUS_ONE = BLOB_RANDOM_VALID1 + b"\x00"
-BLOB_INVALID_LENGTH_MINUS_ONE = BLOB_RANDOM_VALID1[:-1]
-
-VALID_BLOBS = [BLOB_ALL_ZEROS, BLOB_ALL_TWOS, BLOB_RANDOM_VALID1, BLOB_RANDOM_VALID2,
-               BLOB_RANDOM_VALID3, BLOB_ALL_MODULUS_MINUS_ONE, BLOB_ALMOST_ZERO]
-INVALID_BLOBS = [BLOB_INVALID, BLOB_INVALID_CLOSE, BLOB_INVALID_LENGTH_PLUS_ONE, BLOB_INVALID_LENGTH_MINUS_ONE]
-
-FE_VALID1 = field_element_bytes(0)
-FE_VALID2 = field_element_bytes(1)
-FE_VALID3 = field_element_bytes(2)
-FE_VALID4 = field_element_bytes(pow(5, 1235, spec.BLS_MODULUS))
-FE_VALID5 = field_element_bytes(spec.BLS_MODULUS - 1)
-FE_VALID6 = field_element_bytes(spec.compute_roots_of_unity(spec.FIELD_ELEMENTS_PER_BLOB)[1])
-VALID_FIELD_ELEMENTS = [FE_VALID1, FE_VALID2, FE_VALID3, FE_VALID4, FE_VALID5, FE_VALID6]
-
-FE_INVALID_EQUAL_TO_MODULUS = field_element_bytes_unchecked(spec.BLS_MODULUS)
-FE_INVALID_MODULUS_PLUS_ONE = field_element_bytes_unchecked(spec.BLS_MODULUS + 1)
-FE_INVALID_UINT256_MAX = field_element_bytes_unchecked(2**256 - 1)
-FE_INVALID_UINT256_MID = field_element_bytes_unchecked(2**256 - 2**128)
-FE_INVALID_LENGTH_PLUS_ONE = VALID_FIELD_ELEMENTS[0] + b"\x00"
-FE_INVALID_LENGTH_MINUS_ONE = VALID_FIELD_ELEMENTS[0][:-1]
-INVALID_FIELD_ELEMENTS = [FE_INVALID_EQUAL_TO_MODULUS, FE_INVALID_MODULUS_PLUS_ONE,
-                          FE_INVALID_UINT256_MAX, FE_INVALID_UINT256_MID,
-                          FE_INVALID_LENGTH_PLUS_ONE, FE_INVALID_LENGTH_MINUS_ONE]
-
-
-def hash(x):
-    return sha256(x).digest()
-
-
-def int_to_hex(n: int, byte_length: int = None) -> str:
-    byte_value = int_to_big_endian(n)
-    if byte_length:
-        byte_value = byte_value.rjust(byte_length, b'\x00')
-    return encode_hex(byte_value)
-
+###############################################################################
+# Test cases for blob_to_kzg_commitment
+###############################################################################
 
 def case01_blob_to_kzg_commitment():
     # Valid cases
@@ -137,6 +41,10 @@ def case01_blob_to_kzg_commitment():
             'output': None
         }
 
+
+###############################################################################
+# Test cases for compute_kzg_proof
+###############################################################################
 
 def case02_compute_kzg_proof():
     # Valid cases
@@ -178,6 +86,10 @@ def case02_compute_kzg_proof():
             'output': None
         }
 
+
+###############################################################################
+# Test cases for verify_kzg_proof
+###############################################################################
 
 def case03_verify_kzg_proof():
     # Valid cases
@@ -341,6 +253,10 @@ def case03_verify_kzg_proof():
         }
 
 
+###############################################################################
+# Test cases for compute_blob_kzg_proof
+###############################################################################
+
 def case04_compute_blob_kzg_proof():
     # Valid cases
     for blob in VALID_BLOBS:
@@ -381,6 +297,10 @@ def case04_compute_blob_kzg_proof():
             'output': None
         }
 
+
+###############################################################################
+# Test cases for verify_blob_kzg_proof
+###############################################################################
 
 def case05_verify_blob_kzg_proof():
     # Valid cases
@@ -502,6 +422,10 @@ def case05_verify_blob_kzg_proof():
             'output': None
         }
 
+
+###############################################################################
+# Test cases for verify_blob_kzg_proof_batch
+###############################################################################
 
 def case06_verify_blob_kzg_proof_batch():
     # Valid cases
@@ -626,6 +550,10 @@ def case06_verify_blob_kzg_proof_batch():
         'output': None
     }
 
+
+###############################################################################
+# Main logic
+###############################################################################
 
 def create_provider(fork_name: SpecForkName,
                     handler_name: str,
