+++ tests/generators/kzg_7594/main.py
@@ -2,123 +2,16 @@
 KZG test vectors generator for EIP-7594
 """
 
-from hashlib import sha256
 from typing import Tuple, Iterable, Any, Callable, Dict
 
 from eth_utils import encode_hex
 
-from eth2spec.utils import bls
+from eth2spec.eip7594 import spec
+from eth2spec.gen_helpers.gen_base import gen_runner, gen_typing
 from eth2spec.test.helpers.constants import EIP7594
 from eth2spec.test.helpers.typing import SpecForkName
-from eth2spec.gen_helpers.gen_base import gen_runner, gen_typing
-from eth2spec.eip7594 import spec
-
-
-###############################################################################
-# Helper functions
-###############################################################################
-
-def expect_exception(func, *args):
-    try:
-        func(*args)
-    except Exception:
-        pass
-    else:
-        raise Exception("should have raised exception")
-
-
-def bls_add_one(x):
-    """
-    Adds "one" (actually bls.G1()) to a compressed group element.
-    Useful to compute definitely incorrect proofs.
-    """
-    return bls.G1_to_bytes48(
-        bls.add(bls.bytes48_to_G1(x), bls.G1())
-    )
-
-
-def hash(x):
-    return sha256(x).digest()
-
-
-def make_id(*args):
-    values_str = "_".join(str(arg) for arg in args)
-    return hash(bytes(values_str, "utf-8"))[:8].hex()
-
-
-def field_element_bytes(x):
-    return int.to_bytes(x % spec.BLS_MODULUS, 32, spec.KZG_ENDIANNESS)
-
-
-def field_element_bytes_unchecked(x):
-    return int.to_bytes(x, 32, spec.KZG_ENDIANNESS)
-
-
-def encode_hex_list(a):
-    return [encode_hex(x) for x in a]
-
-
-###############################################################################
-# Global variables
-###############################################################################
-
-BLS_MODULUS_BYTES = spec.BLS_MODULUS.to_bytes(32, spec.KZG_ENDIANNESS)
-
-# Blobs
-
-BLOB_ALL_ZEROS = spec.Blob()
-BLOB_ALL_TWOS = spec.Blob(b''.join([field_element_bytes(2) for n in range(4096)]))
-BLOB_RANDOM_VALID1 = spec.Blob(b''.join([field_element_bytes(pow(2, n + 256, spec.BLS_MODULUS)) for n in range(4096)]))
-BLOB_RANDOM_VALID2 = spec.Blob(b''.join([field_element_bytes(pow(3, n + 256, spec.BLS_MODULUS)) for n in range(4096)]))
-BLOB_RANDOM_VALID3 = spec.Blob(b''.join([field_element_bytes(pow(5, n + 256, spec.BLS_MODULUS)) for n in range(4096)]))
-BLOB_ALL_MODULUS_MINUS_ONE = spec.Blob(b''.join([field_element_bytes(spec.BLS_MODULUS - 1) for n in range(4096)]))
-BLOB_ALMOST_ZERO = spec.Blob(b''.join([field_element_bytes(1 if n == 3211 else 0) for n in range(4096)]))
-
-BLOB_INVALID = spec.Blob(b'\xFF' * 4096 * 32)
-BLOB_INVALID_CLOSE = spec.Blob(b''.join(
-    [BLS_MODULUS_BYTES if n == 2111 else field_element_bytes(0) for n in range(4096)]
-))
-BLOB_INVALID_LENGTH_PLUS_ONE = BLOB_RANDOM_VALID1 + b"\x00"
-BLOB_INVALID_LENGTH_MINUS_ONE = BLOB_RANDOM_VALID1[:-1]
-
-VALID_BLOBS = [BLOB_ALL_ZEROS, BLOB_ALL_TWOS, BLOB_RANDOM_VALID1, BLOB_RANDOM_VALID2,
-               BLOB_RANDOM_VALID3, BLOB_ALL_MODULUS_MINUS_ONE, BLOB_ALMOST_ZERO]
-INVALID_BLOBS = [BLOB_INVALID, BLOB_INVALID_CLOSE, BLOB_INVALID_LENGTH_PLUS_ONE, BLOB_INVALID_LENGTH_MINUS_ONE]
-
-# Individual Cells
-
-CELL_RANDOM_VALID1 = b"".join([field_element_bytes(pow(2, n + 256, spec.BLS_MODULUS))
-                               for n in range(spec.FIELD_ELEMENTS_PER_CELL)])
-CELL_RANDOM_VALID2 = b"".join([field_element_bytes(pow(3, n + 256, spec.BLS_MODULUS))
-                               for n in range(spec.FIELD_ELEMENTS_PER_CELL)])
-CELL_RANDOM_VALID3 = b"".join([field_element_bytes(pow(5, n + 256, spec.BLS_MODULUS))
-                               for n in range(spec.FIELD_ELEMENTS_PER_CELL)])
-
-CELL_ALL_MAX_VALUE = b"".join([field_element_bytes_unchecked(2 ** 256 - 1)
-                               for n in range(spec.FIELD_ELEMENTS_PER_CELL)])
-CELL_ONE_INVALID_FIELD = b"".join([field_element_bytes_unchecked(spec.BLS_MODULUS)
-                                   if n == 7 else field_element_bytes(0)
-                                   for n in range(spec.FIELD_ELEMENTS_PER_CELL)])
-
-VALID_INDIVIDUAL_RANDOM_CELL_BYTES = [CELL_RANDOM_VALID1, CELL_RANDOM_VALID2, CELL_RANDOM_VALID3]
-INVALID_INDIVIDUAL_CELL_BYTES = [CELL_ALL_MAX_VALUE, CELL_ONE_INVALID_FIELD]
-
-# Cells & Proofs
-
-VALID_CELLS_AND_PROOFS = []  # Saved in case02_compute_cells_and_proofs
-VALID_COMMITMENTS = [spec.blob_to_kzg_commitment(blob) for blob in VALID_BLOBS]
-
-# Points
-
-G1 = bls.G1_to_bytes48(bls.G1())
-G1_INVALID_TOO_FEW_BYTES = G1[:-1]
-G1_INVALID_TOO_MANY_BYTES = G1 + b"\x00"
-G1_INVALID_P1_NOT_IN_G1 = bytes.fromhex("8123456789abcdef0123456789abcdef0123456789abcdef" +
-                                        "0123456789abcdef0123456789abcdef0123456789abcdef")
-G1_INVALID_P1_NOT_ON_CURVE = bytes.fromhex("8123456789abcdef0123456789abcdef0123456789abcdef" +
-                                           "0123456789abcdef0123456789abcdef0123456789abcde0")
-INVALID_G1_POINTS = [G1_INVALID_TOO_FEW_BYTES, G1_INVALID_TOO_MANY_BYTES,
-                     G1_INVALID_P1_NOT_IN_G1, G1_INVALID_P1_NOT_ON_CURVE]
+from eth2spec.test.utils.kzg_tests import *
+from eth2spec.utils import bls
 
 
 ###############################################################################
